Put host port in container's subgraph rather than in the outside
At the beginning, I thought it makes sense to put host ports outside the network's subgraphs to represent the machine itself, but it leads to very long edges and makes the graph hard to read. If you have the legend, it is not confusing to see the host ports near the container.